+++ src/lib/fragments/AgGrid.react.js
@@ -17,6 +17,7 @@ import {
 import {
     propTypes as _propTypes,
     defaultProps as _defaultProps,
+    apiGetters,
 } from '../components/AgGrid.react';
 import {
     COLUMN_DANGEROUS_FUNCTIONS,
@@ -73,6 +74,9 @@ const xssMessage = (context) => {
 const NO_CONVERT_PROPS = [...PASSTHRU_PROPS, ...PROPS_NOT_FOR_AG_GRID];
 
 const agGridRefs = {};
+
+apiGetters.getApi = (id) => agGridRefs[stringifyId(id)]?.api;
+apiGetters.getColumnApi = (id) => agGridRefs[stringifyId(id)]?.columnApi;
 
 const eventBus = {
     listeners: {},