+++ src/lib/fragments/AgGrid.react.js
@@ -457,11 +457,7 @@ export default class DashAgGrid extends Component {
         const filterModel = this.state.gridApi.getFilterModel();
         const propsToSet = {filterModel};
         if (rowModelType === 'clientSide') {
-            const virtualRowData = [];
-            this.state.gridApi.forEachNodeAfterFilterAndSort((node) => {
-                virtualRowData.push(node.data);
-            });
-            propsToSet.virtualRowData = virtualRowData;
+            propsToSet.virtualRowData = this.virtualRowData();
         }
 
         setProps(propsToSet);
@@ -469,23 +465,33 @@ export default class DashAgGrid extends Component {
 
     getRowData() {
         const newRowData = [];
-        this.state.gridApi.forEachNode((node) => {
+        this.state.gridApi.forEachLeafNode((node) => {
             newRowData.push(node.data);
         });
         return newRowData;
     }
 
-    syncRowData() {
-        const {rowData, setProps, rowModelType} = this.props;
-        if (rowData) {
-            const virtualRowData = [];
-            if (rowModelType === 'clientSide') {
-                this.state.gridApi.forEachNodeAfterFilterAndSort((node) => {
+    virtualRowData() {
+        const {rowModelType} = this.props;
+        const {gridApi} = this.state;
+        const virtualRowData = [];
+        if (rowModelType === 'clientSide' && gridApi) {
+            gridApi.forEachNodeAfterFilterAndSort((node) => {
+                if (node.data) {
                     virtualRowData.push(node.data);
-                });
-            }
-            if (rowData !== this.getRowData()) {
-                setProps({rowData: this.getRowData(), virtualRowData});
+                }
+            });
+        }
+        return virtualRowData;
+    }
+
+    syncRowData() {
+        const {rowData, setProps} = this.props;
+        if (rowData) {
+            const virtualRowData = this.virtualRowData();
+            const newRowData = this.getRowData();
+            if (rowData !== newRowData) {
+                setProps({rowData: newRowData, virtualRowData});
             } else {
                 setProps({virtualRowData});
             }
@@ -496,12 +502,7 @@ export default class DashAgGrid extends Component {
         const {setProps, rowModelType} = this.props;
         const propsToSet = {};
         if (rowModelType === 'clientSide') {
-            const virtualRowData = [];
-            this.state.gridApi.forEachNodeAfterFilterAndSort((node) => {
-                virtualRowData.push(node.data);
-            });
-
-            propsToSet.virtualRowData = virtualRowData;
+            propsToSet.virtualRowData = this.virtualRowData();
         }
         propsToSet.columnState = JSON.parse(
             JSON.stringify(this.state.gridColumnApi.getColumnState())
@@ -779,10 +780,7 @@ export default class DashAgGrid extends Component {
             this.setSelection(selectedRows);
 
             if (rowData && rowModelType === 'clientSide') {
-                const virtualRowData = [];
-                gridApi.forEachNodeAfterFilterAndSort((node) => {
-                    virtualRowData.push(node.data);
-                });
+                const virtualRowData = this.virtualRowData();
 
                 setProps({virtualRowData});
             }
@@ -902,12 +900,7 @@ export default class DashAgGrid extends Component {
         node,
     }) {
         const timestamp = Date.now();
-        const virtualRowData = [];
-        if (this.props.rowModelType === 'clientSide' && this.state.gridApi) {
-            this.state.gridApi.forEachNodeAfterFilterAndSort((node) => {
-                virtualRowData.push(node.data);
-            });
-        }
+        const virtualRowData = this.virtualRowData();
         this.props.setProps({
             cellValueChanged: {
                 rowIndex,
@@ -1201,8 +1194,8 @@ export default class DashAgGrid extends Component {
         if (reset) {
             this.props.setProps({
                 deleteSelectedRows: false,
-                rowData: this.getRowData(),
             });
+            this.syncRowData();
         }
     }
     // end event actions
@@ -1243,8 +1236,8 @@ export default class DashAgGrid extends Component {
                 this.applyRowTransaction(data);
                 this.props.setProps({
                     rowTransaction: null,
-                    rowData: this.getRowData(),
                 });
+                this.syncRowData();
             } else {
                 this.setState({
                     rowTransaction: rowTransaction