+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Links "DE#nnn" prior to version 2.0 point to the Dash Enterprise closed-source D
 ## [UNRELEASED]
 ### Added
 - [#199](https://github.com/plotly/dash-ag-grid/pull/199) Add `scrollTo` prop which allows scrolling to rows and columns.
+- [#209](https://github.com/plotly/dash-ag-grid/pull/209)
+  - Adding `getApi` and `getColumnApi` to `dash_ag_grid` namespace to allow for JS functions to call the grid's API directly
 
 ## [2.1.0] - 2023-06-02