+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to `dash-ag-grid` will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 Links "DE#nnn" prior to version 2.0 point to the Dash Enterprise closed-source Dash AG Grid repo
 
+## [Unreleased] -
+### Added
+- [#212](https://github.com/plotly/dash-ag-grid/pull/212) Async function for `getApiAsync` and `getColumnApiAsync` for use with grid initializing 
+
 ## [2.2.0] - 2023-06-20
 
 ### Added