+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Links "DE#nnn" prior to version 2.0 point to the Dash Enterprise closed-source D
 - [#212](https://github.com/plotly/dash-ag-grid/pull/212) Async function for `getApiAsync` and `getColumnApiAsync` for use with grid initializing 
 
 ### Fixed
-- [#226](https://github.com/plotly/dash-ag-grid/pull/226) Fixed issue when using grouped rows with `rowData` and `virtualRowData` populating incorrectly.
+- [#226](https://github.com/plotly/dash-ag-grid/pull/226) Fixed issue when using grouped rows with `rowData` and `virtualRowData` populating incorrectly. [#215](https://github.com/plotly/dash-ag-grid/issues/215)
 
 ## [2.2.0] - 2023-06-20